requirements.txt
anthropic
google-cloud-bigquery
google-genai
hdbscan
openai
python-dotenv
requests
scikit-learn
umap-learn
